Struct cpp_to_rust_generator::config::Config

pub struct Config { /* fields omitted */ }

The starting point of cpp_to_rust API. Create a Config object, set its properties, add custom functions if necessary, and start the processing with Config::exec.

Methods

impl Config

fn new<P1: Into<PathBuf>, P2: Into<PathBuf>>(output_dir_path: P1,
cache_dir_path: P2,
crate_properties: CrateProperties)
-> Config

Creates a Config. crate_properties are used in Cargo.toml of the generated crate. output_dir_path will contain the generated crate. cache_dir_path will be used for cache, temporary files and inter-library information files.

fn is_completed(&self) -> bool

Returns true if a completion marker exists in the cache directory, indicating that processing of the library in this cache directory was completed before. Note that the marker is not created if write_cache was set to false. The marker will only be used if cache usage is set to CacheUsage::Full.

fn completed_marker_path(&self) -> PathBuf

Returns path to the completion marker file indicating that processing of the library in this cache directory was completed before. Note that the marker is not created if write_cache was set to false. The marker will only be used if cache usage is set to CacheUsage::Full.

fn set_cache_usage(&mut self, value: CacheUsage)

Defines how cached data is used in repeated runs of the generator.

fn set_crate_template_path<P: Into<PathBuf>>(&mut self, path: P)

Sets the directory containing additional files for the crate. Any files and directories found in the crate template will be copied to the generated crate's directory, although some of them (such as Cargo.toml) may be overwritten with the generates files. It's common to put tests and examples subdirectories in the crate template so that cargo recognizes them automatically in the generated crate.

If you want to add some extra code to the generated modules, put src/module_name.rs file in the crate template and add include_generated!(); line in the file. This line will be replaced with the generated content. You can also add extra modules as separate files, but you'll also need to create src/lib.rs in the crate template and declare new module in it using [pub] mod module_name;. Use include_generated!(); in src/lib.rs to include declaration of automatically generated modules.

If the crate template contains rustfmt.toml file, it's used to format the generated Rust code instead of the default rustfmt.toml.

Creating crate template is optional. The generator can make a crate without a template.

fn set_dependency_cache_paths(&mut self, paths: Vec<PathBuf>)

Sets list of paths to cache directories of processed dependencies. The generator will integrate API of the current library with its dependencies and re-use their types.

fn add_cpp_parser_blocked_name<P: Into<String>>(&mut self, lib: P)

Adds a C++ identifier that should be skipped by the C++ parser. Identifier can contain namespaces and nested classes, with :: separator (like in C++ identifiers). Identifier may refer to a method, a class, a enum or a namespace. All entities inside blacklisted entity (e.g. the methods of a blocked class or the contents of a blocked namespace) will also be skipped. All class methods with names matching the blocked name will be skipped, regardless of class name.

fn add_cpp_parser_blocked_names<Item, Iter>(&mut self, items: Iter) where Item: Into<String>, Iter: IntoIterator<Item=Item>

Adds multiple blocked names. See Config::add_cpp_parser_blocked_name.

fn add_cpp_parser_argument<P: Into<String>>(&mut self, lib: P)

Adds a command line argument for clang C++ parser.

Note that this value is not used when building the wrapper library. Use Config::cpp_build_config_mut or a similar method to configure building the wrapper library.

fn add_cpp_parser_arguments<Item, Iter>(&mut self, items: Iter) where Item: Into<String>, Iter: IntoIterator<Item=Item>

Adds multiple command line arguments for clang C++ parser. See Config::add_cpp_parser_argument.

fn add_include_path<P: Into<PathBuf>>(&mut self, path: P)

Adds path to an include directory. It's supplied to the C++ parser via -I option.

Note that this value is not used when building the wrapper library. Use Config::cpp_build_config_mut or a similar method to configure building the wrapper library.

fn add_framework_path<P: Into<PathBuf>>(&mut self, path: P)

Adds path to a framework directory (OS X specific). It's supplied to the C++ parser via -F option.

Note that this value is not used when building the wrapper library. Use Config::cpp_build_config_mut or a similar method to configure building the wrapper library.

fn add_target_include_path<P: Into<PathBuf>>(&mut self, path: P)

Adds path to an include directory or an include file of the target library. Any C++ types and methods will be parsed and used only if they are declared within one of files or directories added with this method.

If no target include paths are added, all types and methods will be used. Most libraries include system headers and other libraries' header files, so this mode is often unwanted.

fn add_include_directive<P: Into<PathBuf>>(&mut self, path: P)

Adds an include directive. Each directive will be added as #include <path> to the input file for the C++ parser. File name only paths or relative paths should be used in this method.

fn add_cpp_ffi_generator_filter<F>(&mut self, f: F) where F: Fn(&CppMethod) -> Result<bool> + 'static

Adds a custom function that decides whether a C++ method should be added to the C++ wrapper library. For each C++ method, each function will be run once. Filters are executed in the same order they were added.

Interpetation of the function's output:

Err indicates an unexpected failure and terminates the processing.

Ok(true) allows to continue processing of the method. If all functions return Ok(true), the method is accepted.

Ok(false) blocks the method. Remaining filter functions are not run on this method.

fn add_cpp_data_filter<F>(&mut self, f: F) where F: Fn(&mut CppData) -> Result<()> + 'static

Adds a custom function that visits &mut CppData and can perform any changes in the output of the C++ parser. Filters are executed in the same order they were added. If the function returns Err, the processing is terminated.

fn set_type_allocation_place<S: Into<String>>(&mut self,
place: CppTypeAllocationPlace,
type_name: S)

Overrides automatic selection of type allocation place for type_name and uses place instead. See CppTypeAllocationPlace for more information.

fn set_types_allocation_place<SI, S>(&mut self,
place: CppTypeAllocationPlace,
types: SI) where SI: IntoIterator<Item=S>, S: Into<String>

Overrides automatic selection of type allocation place for types and uses place instead. See also Config::set_type_allocation_place.

fn set_debug_logging_config(&mut self, config: DebugLoggingConfig)

Changes how debug logs are handled. See DebugLoggingConfig for more information.

fn set_quiet_mode(&mut self, quiet_mode: bool)

Sets quiet mode. In quiet mode status messages and debug logs are redirected to log subdirectory of the cache directory. Only error messages are always written to stderr. Quiet mode is disabled by default.

fn set_write_cache(&mut self, write_cache: bool)

Sets writing to cache mode. If enabled, result of processing is saved to extra files in cache directory. The main use of these files are loading dependency data. They are also used for speeding up repeated runs of the generator on the same library.

Writing to cache is enabled by default. If the library is not intended to be used as a dependency when running the generator on another library, it's safe to disable writing to cache in order to speed up the generator.

fn set_cpp_build_config(&mut self, cpp_build_config: CppBuildConfig)

Sets CppBuildConfig value that will be passed to the build script of the generated crate.

fn cpp_build_config_mut(&mut self) -> &mut CppBuildConfig

Allows to change CppBuildConfig value that will be passed to the build script of the generated crate.

fn set_cpp_lib_version<S: Into<String>>(&mut self, version: S)

fn cpp_lib_version(&self) -> Option<&str>

fn exec(self) -> Result<()>

Starts execution of the generator. This function will print the necessary build script output to stdout. It also displays some debugging output that can be made visible by running cargo commands with -vv option.

The result of this function must be checked. You can use ::errors::fancy_unwrap to check the result and display additional error information.

fn cache_usage(&self) -> &CacheUsage

Returns value set by Config::set_cache_usage.

fn crate_properties(&self) -> &CrateProperties

Returns crate properties passed to Config::new.

fn output_dir_path(&self) -> &PathBuf

Returns path to the output directory passed to Config::new.

fn cache_dir_path(&self) -> &PathBuf

Returns path to the cache directory passed to Config::new.

fn crate_template_path(&self) -> Option<&PathBuf>

Returns value set by Config::set_crate_template_path.

fn dependency_cache_paths(&self) -> &[PathBuf]

Returns value set by Config::set_dependency_cache_paths.

fn cpp_parser_blocked_names(&self) -> &[String]

Returns names added with Config::add_cpp_parser_blocked_name and similar methods.

fn cpp_parser_arguments(&self) -> &[String]

Returns names added with Config::add_cpp_parser_argument and similar methods.

fn include_paths(&self) -> &[PathBuf]

Returns values added by Config::add_include_path.

fn framework_paths(&self) -> &[PathBuf]

Returns values added by Config::add_framework_path.

fn target_include_paths(&self) -> &[PathBuf]

Returns values added by Config::add_target_include_path.

fn include_directives(&self) -> &[PathBuf]

Returns values added by Config::add_include_directive.

fn cpp_ffi_generator_filters(&self) -> Vec<&Box<CppFfiGeneratorFilterFn>>

Returns values added by Config::add_cpp_ffi_generator_filter.

fn cpp_data_filters(&self) -> Vec<&Box<CppDataFilterFn>>

Returns values added by Config::add_cpp_data_filter.

fn cpp_build_config(&self) -> &CppBuildConfig

Returns current CppBuildConfig value.

fn type_allocation_places(&self) -> &HashMap<String, CppTypeAllocationPlace>

Returns values added by Config::set_type_allocation_place. Keys of the hash map are names of C++ types.

fn set_write_dependencies_local_paths(&mut self, value: bool)

If value is true, the generated Cargo.toml will specify both versions and local paths of all dependencies. If value is false, only version will be specified, so publishing all dependencies would be required to build the crate.

fn write_dependencies_local_paths(&self) -> bool

Returns value set by Config::set_write_dependencies_local_paths.

fn debug_logging_config(&self) -> &DebugLoggingConfig

Returns value set by Config::set_debug_logging_config.

fn quiet_mode(&self) -> bool

Returns value set by Config::set_quiet_mode.

fn write_cache(&self) -> bool

Returns value set by Config::set_write_cache.