Struct cpp_to_rust_generator::config::CrateProperties

pub struct CrateProperties { /* fields omitted */ }

Information about the crate being generated. Most of information in this object will be used in the output Cargo.toml.

Methods

impl CrateProperties

fn new<S1: Into<String>, S2: Into<String>>(name: S1,
version: S2)
-> CrateProperties

Creates a new object with name and version.

fn add_author<S: Into<String>>(&mut self, author: S)

Adds an author for authors field of Cargo.toml.

fn set_links_attribute<S: Into<String>>(&mut self, links: S)

Sets value for links field of Cargo.toml. Native name of the C++ library should be used here.

fn add_dependency<S1: Into<String>, S2: Into<String>>(&mut self,
name: S1,
version: S2,
local_path: Option<PathBuf>)

Adds an extra non-cpp_to_rust-based dependency with name, version and optionally local_path.

fn add_build_dependency<S1: Into<String>, S2: Into<String>>(&mut self,
name: S1,
version: S2,
local_path: Option<PathBuf>)

Adds an extra build dependency with name, version and optionally local_path.

fn remove_default_dependencies(&mut self)

Removes default dependencies from output Cargo.toml. Default dependencies are libc, cpp_utils and crates added using Config::set_dependency_cache_paths.

fn remove_default_build_dependencies(&mut self)

Removes default build dependencies from output Cargo.toml. Default build dependency is cpp_to_rust_build_tools.

fn name(&self) -> &String

Name of the crate

fn version(&self) -> &String

Version of the crate

fn authors(&self) -> &Vec<String>

Authors of the crate

fn links_attribute(&self) -> Option<&String>

links property of the crate

fn dependencies(&self) -> &Vec<CrateDependency>

Extra non-cpp_to_rust-based dependencies of the crate

fn build_dependencies(&self) -> &Vec<CrateDependency>

Extra build dependencies of the crate

fn should_remove_default_dependencies(&self) -> bool

Returns true if default dependencies were removed.

fn should_remove_default_build_dependencies(&self) -> bool

Returns true if default build dependencies were removed.