Module cpp_to_rust_generator::cpp_data

Types for handling information about C++ library APIs.

Structs

CppBaseSpecifier
Item of base class list in a class declaration

CppClassField
Member field of a C++ class declaration

CppClassUsingDirective
A "using" directive inside a class definition, indicating that the class should inherite a certain method of a base class.

CppData
C++ parser output

CppEnumValue
One item of a C++ enum declaration

CppOriginLocation
Location of a C++ type's definition in header files.

CppTemplateInstantiation
Information about a C++ template class instantiation.

CppTemplateInstantiations
List of template instantiations of a template class.

CppTypeData
Information about a C++ type declaration

CppTypeDoc
C++ documentation for a type

TemplateArgumentsDeclaration
Information about template arguments of a C++ class type

Enums

CppTypeAllocationPlace
Type allocation place of a C++ type.

CppTypeKind
Information about a C++ type declaration

CppVisibility
Visibility of a C++ entity. Defaults to Public for entities that can't have visibility (like free functions)

Functions

create_cast_method
Convenience function to create CppMethod object for static_cast or dynamic_cast from type from to type to. See CppMethod's documentation for more information about is_unsafe_static_cast and is_direct_static_cast.