Struct cpp_to_rust_generator::cpp_data::CppData

pub struct CppData {
    pub types: Vec<CppTypeData>,
    pub methods: Vec<CppMethod>,
    pub template_instantiations: Vec<CppTemplateInstantiations>,
    pub signal_argument_types: Vec<Vec<CppType>>,
    pub type_allocation_places: HashMap<String, CppTypeAllocationPlace>,
    pub dependencies: Vec<CppData>,
}

C++ parser output

Fields
types: Vec<CppTypeData>

List of found type declarations

methods: Vec<CppMethod>

List of found methods

template_instantiations: Vec<CppTemplateInstantiations>

List of found template instantiations. Key is name of the template class, value is list of instantiations.

signal_argument_types: Vec<Vec<CppType>>

List of all argument types used by signals, including variations with omitted arguments, but excluding argument types from dependencies.

type_allocation_places: HashMap<String, CppTypeAllocationPlace>

List of selected (automatically or in configuration) type allocation places for all class types.

dependencies: Vec<CppData>

Data of dependencies

Methods

impl CppData

fn find_type_info<F>(&self, f: F) -> Option<&CppTypeData> where F: Fn(&&CppTypeData) -> bool

Search for a CppTypeData object in this CppData and all dependencies.

fn is_polymorphic_type(&self, name: &str) -> bool

Returns true if C++ type name is polymorphic, i.e. has at least one virtual function.

fn ensure_explicit_destructors(&mut self) -> Result<()>

Adds destructors for every class that does not have explicitly defined destructor, allowing to create wrappings for destructors implicitly available in C++.

fn add_inherited_methods(&mut self) -> Result<()>

Adds methods of derived classes inherited from base classes. A method will not be added if there is a method with the same name in the derived class. Constructors, destructors and assignment operators are also not added. This reflects C++'s method inheritance rules.

fn generate_methods_with_omitted_args(&mut self)

Generates duplicate methods with fewer arguments for C++ methods with default argument values.

fn all_include_files(&self) -> Result<HashSet<String>>

Returns all include files found within this CppData (excluding dependencies).

fn is_template_class(&self, name: &str) -> bool

Checks if specified class is a template class.

fn has_virtual_destructor(&self, class_name: &str) -> bool

Checks if specified class has virtual destructor (own or inherited).

fn has_virtual_methods(&self, class_name: &str) -> bool

Checks if specified class has any virtual methods (own or inherited).

fn has_public_destructor(&self, class_type: &CppTypeClassBase) -> bool

Checks if specified class has public destructor.

fn has_pure_virtual_methods(&self, class_name: &str) -> bool

Checks if specified class has any pure virtual methods.

fn add_field_accessors(&mut self) -> Result<()>

Adds fictional getter and setter methods for each known public field of each class.

fn inherits(&self, class_name: &str, base_name: &str) -> bool

Checks if class_name types inherits base_name type directly or indirectly.

fn choose_allocation_places(&mut self,
overrides: &HashMap<String, CppTypeAllocationPlace>)
-> Result<()>

Detects the preferred type allocation place for each type based on API of all known methods. Keys of overrides are C++ type names. If overrides contains type allocation place for a type, it's used instead of the place that would be automatically selected.

fn type_allocation_place(&self,
class_name: &str)
-> Result<CppTypeAllocationPlace>

Returns selected type allocation place for type class_name.

fn post_process(&mut self) -> Result<()>

Performs data conversion to make it more suitable for further wrapper generation.