Enum cpp_to_rust_generator::cpp_method::CppOperator

pub enum CppOperator {
    Conversion(CppType),
    Assignment,
    Addition,
    Subtraction,
    UnaryPlus,
    UnaryMinus,
    Multiplication,
    Division,
    Modulo,
    PrefixIncrement,
    PostfixIncrement,
    PrefixDecrement,
    PostfixDecrement,
    EqualTo,
    NotEqualTo,
    GreaterThan,
    LessThan,
    GreaterThanOrEqualTo,
    LessThanOrEqualTo,
    LogicalNot,
    LogicalAnd,
    LogicalOr,
    BitwiseNot,
    BitwiseAnd,
    BitwiseOr,
    BitwiseXor,
    BitwiseLeftShift,
    BitwiseRightShift,
    AdditionAssignment,
    SubtractionAssignment,
    MultiplicationAssignment,
    DivisionAssignment,
    ModuloAssignment,
    BitwiseAndAssignment,
    BitwiseOrAssignment,
    BitwiseXorAssignment,
    BitwiseLeftShiftAssignment,
    BitwiseRightShiftAssignment,
    Subscript,
    Indirection,
    AddressOf,
    StructureDereference,
    PointerToMember,
    FunctionCall,
    Comma,
    New,
    NewArray,
    Delete,
    DeleteArray,
}

Available types of C++ operators

Variants
Conversion(CppType)

(type) a

Assignment

a = b

Addition

a + b

Subtraction

a - b

UnaryPlus

+a

UnaryMinus

-a

Multiplication

a * b

Division

a / b

Modulo

a % b

PrefixIncrement

++a

PostfixIncrement

a++

PrefixDecrement

--a

PostfixDecrement

a--

EqualTo

a == b

NotEqualTo

a != b

GreaterThan

a > b

LessThan

a < b

GreaterThanOrEqualTo

a >= b

LessThanOrEqualTo

a <= b

LogicalNot

!a

LogicalAnd

a && b

LogicalOr

a || b

BitwiseNot

~a

BitwiseAnd

a & b

BitwiseOr

a | b

BitwiseXor

a ^ b

BitwiseLeftShift

a << b

BitwiseRightShift

a >> b

AdditionAssignment

a += b

SubtractionAssignment

a -= b

MultiplicationAssignment

a *= b

DivisionAssignment

a /= b

ModuloAssignment

a %= b

BitwiseAndAssignment

a &= b

BitwiseOrAssignment

a |= b

BitwiseXorAssignment

a = b

BitwiseLeftShiftAssignment

a <<= b

BitwiseRightShiftAssignment

a >>= b

Subscript

a[b]

Indirection

*a

AddressOf

&a

StructureDereference

a->b

PointerToMember

a->*b

FunctionCall

a(a1, a2)

Comma

a, b

New

new type

NewArray

new type[n]

Delete

delete a

DeleteArray

delete[] a

Methods

impl CppOperator

fn info(&self) -> CppOperatorInfo

Reports information about this operator

fn c_name(&self) -> Result<&'static str>

Returns alphanumeric identifier for this operator used to name FFI functions.

fn all() -> Vec<CppOperator>

Returns all existing operator kinds except for conversion operator which includes an arbitrary C++ type.