Enum cpp_to_rust_generator::cpp_type::CppBuiltInNumericType

pub enum CppBuiltInNumericType {
    Bool,
    Char,
    SChar,
    UChar,
    WChar,
    Char16,
    Char32,
    Short,
    UShort,
    Int,
    UInt,
    Long,
    ULong,
    LongLong,
    ULongLong,
    Int128,
    UInt128,
    Float,
    Double,
    LongDouble,
}

Available built-in C++ numeric types. All these types have corresponding clang::TypeKind values (except for CharS and CharU which map to CppBuiltInNumericType::Char)

Variants
BoolCharSCharUCharWCharChar16Char32ShortUShortIntUIntLongULongLongLongULongLongInt128UInt128FloatDoubleLongDouble
Methods

impl CppBuiltInNumericType

fn to_cpp_code(&self) -> &'static str

Returns C++ code representing this type.

fn is_float(&self) -> bool

Returns true if this type is some sort of floating point type.

fn is_signed_integer(&self) -> bool

Returns true if this type is a signed integer.

fn is_unsigned_integer(&self) -> bool

Returns true if this type is an unsigned integer.

fn is_integer_with_undefined_signedness(&self) -> bool

Returns true if this type is integer but may be signed or unsigned, depending on the platform.

fn all() -> &'static [CppBuiltInNumericType]

Returns all supported types.