Enum cpp_to_rust_generator::cpp_type::CppTypeBase

pub enum CppTypeBase {
    Void,
    BuiltInNumeric(CppBuiltInNumericType),
    SpecificNumeric(CppSpecificNumericType),
    PointerSizedInteger {
        name: String,
        is_signed: bool,
    },
    Enum {
        name: String,
    },
    Class(CppTypeClassBase),
    TemplateParameter {
        nested_level: usize,
        index: usize,
    },
    FunctionPointer(CppFunctionPointerType),
}

Base C++ type. CppType can add indirection and constness to CppTypeBase, but otherwise this enum lists all supported types.

Variants
Void

Void

BuiltInNumeric(CppBuiltInNumericType)

Built-in C++ primitive type, like int

SpecificNumeric(CppSpecificNumericType)

Fixed-size primitive type, like qint64 or int64_t (may be translated to Rust's i64)

PointerSizedInteger

Pointer sized integer, like qintptr (may be translated to Rust's isize)

Fields of PointerSizedInteger

name: String

is_signed: bool

Enum

Enum type

Fields of Enum

name: String

Name, including namespaces and nested classes

Class(CppTypeClassBase)

Class type

TemplateParameter

Template parameter, like "T" anywhere inside QVector<T> declaration

Fields of TemplateParameter

nested_level: usize

Template instantiation level. For example, if there is a template class and a template method in it, the class's template parameters will have level = 0 and the method's template parameters will have level = 1. If only the class or only the method is a template, the level will be 0.

index: usize

Index of the parameter. In QHash<K, V> "K" has index = 0 and "V" has index = 1.

FunctionPointer(CppFunctionPointerType)

Function pointer type

Methods

impl CppTypeBase

fn is_void(&self) -> bool

Returns true if this is void type.

fn is_class(&self) -> bool

Returns true if this is a class type.

fn is_template_parameter(&self) -> bool

Returns true if this is a template parameter.

fn is_function_pointer(&self) -> bool

Returns true if this is a function pointer.

fn is_or_contains_template_parameter(&self) -> bool

Returns true if this is a template parameter or a type that contains any template parameters.

fn to_cpp_code(&self,
function_pointer_inner_text: Option<&str>)
-> Result<String>

Returns C++ code representing this type.

fn caption(&self, strategy: TypeCaptionStrategy) -> Result<String>

Generates alphanumeric representation of self used to generate FFI function names

fn to_cpp_pseudo_code(&self) -> String

Generates string representation of this type for debugging output.