Struct cpp_to_rust_generator::cpp_type::CppType

pub struct CppType {
    pub base: CppTypeBase,
    pub indirection: CppTypeIndirection,
    pub is_const: bool,
    pub is_const2: bool,
}

Information about a C++ type

Fields
base: CppTypeBase

Information about base type

indirection: CppTypeIndirection

Indirection applied to base type

is_const: bool

If the type has no indirection, is_const indicates constness of the type itself (e.g. int vs const int). For types that cannot be const, is_const defaults to false. If the type has one level of indirection, is_const indicates constness of that indirection, i.e. if the pointer or the reference is const. If the type has two levels of indirection, is_const indicates constness of indirection that is applied first.

is_const2: bool

If the type has two levels of indirection, is_const2 indicates constness of indirection that is applied second. In other cases it is false.

Methods

impl CppType

fn void() -> Self

Creates a void type.

fn is_void(&self) -> bool

Returns true if this type is void.

fn to_cpp_code(&self,
function_pointer_inner_text: Option<&str>)
-> Result<String>

Returns C++ code representing this type.

fn to_cpp_pseudo_code(&self) -> String

Returns string representation of this type for debugging output.

fn to_cpp_ffi_type(&self, role: CppTypeRole) -> Result<CppFfiType>

Converts this C++ type to its adaptation for FFI interface, removing all features not supported by C ABI (e.g. references and passing objects by value).

fn caption(&self, strategy: TypeCaptionStrategy) -> Result<String>

Generates alphanumeric representation of self used to generate FFI function names

fn needs_allocation_place_variants(&self) -> bool

Checks if a function with this return type would need to have 2 wrappers with 2 different return value allocation places

fn instantiate(&self,
nested_level1: usize,
template_arguments1: &[CppType])
-> Result<CppType>

Attempts to replace template types at nested_level1 within this type with template_arguments1.

fn is_platform_dependent(&self) -> bool

Returns true if this type is platform dependent. Built-in numeric types that can have different size and/or signedness on different platforms are considered platform dependent. Any types that refer to a platform dependent type are also platform dependent.

Note that most types are platform dependent in "binary" sense, i.e. their size and memory layout may vary, but this function does not address this property.

fn can_be_the_same_as(&self, other_type: &CppType) -> bool

Returns true if self and other_type may theoretically be the same concrete type on some platform. For example, int and long can be the same, but int and unsigned long cannot.